requirements.txt
Django==2.1.3
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==11.0.0
django-debug-toolbar==1.10.1
django-redis==4.10.0
django-storages==1.7.1
mysqlclient==1.3.13
pytz==2018.7
redis==3.0.1
sqlparse==0.2.4
uWSGI==2.0.15
boto3==1.9.71
tqdm==4.28.1
django-celery-results==1.0.4
python-dotenv==0.17.1
numpy==1.20.3
opencv-python==4.5.2.52
scikit-learn==0.24.2
scipy==1.6.3
Pillow==8.2.0
uWSGI==2.0.15
